调整日志输出格式,NetServer增加IO日志输出

+++ mwnet_mt/base/Logging.h
@@ -65,7 +65,7 @@ class Logger
   static LogLevel logLevel();
   static void setLogLevel(LogLevel level);
 
-  typedef void (*OutputFunc)(const char* msg, int len);
+  typedef void (*OutputFunc)(const char* msg, size_t len);
   typedef void (*FlushFunc)();
   static void setOutput(OutputFunc);
   static void setFlush(FlushFunc);
